popover_menus: Extract popovers using tippy in a separate module.
