+++ tools/test-js-with-node
@@ -104,6 +104,7 @@ EXEMPT_FILES = {
     "static/js/pm_list_dom.js",
     "static/js/poll_widget.js",
     "static/js/popovers.js",
+    "static/js/popover_menus.js",
     "static/js/ready.js",
     "static/js/realm_icon.js",
     "static/js/realm_logo.js",
